requirements.txt
Flask==2.3.2
google-cloud-vision==3.4.1
